release: support retrying rollbacks
This commit adds support for retrying rollbacks using
`.spec.rollback.retry` until the configured `.spec.rolback.maxRetries`
count is reached (defaults to 5, setting it to 0 means infinite). The
reset of the counter happens after a successful upgrade or a change
(e.g. chart change, `.spec` change).

This makes it possible to for example recover from transient errors
that were the cause of the rollback, and not a persistent chart or
values failure.